+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 0.2.0:
   * Add `YearMonth::add/sub_years` and `YearMonth::add/sub_months`.
   * Add `Date::add/sub_years` and `Date::add/sub_months`.
   * Add `InvalidDayOfMonth::next/prev_valid()` to get the nearest valid date.
+  * Add `DateResultExt` extension trait for `Result<Date, InvalidDayOfMonth>` to easily round invalid dates.
   * Rename `InvalidDate::InvalidDayForMonth` to `InvalidDate::InvalidDayOfMonth` to match the wrapped struct.
   * Rename `InvalidDayOfYear::day` to `day_of year`.